update
修改消息回调和消息字段

+++ center/src/main/java/center/client/handle/NotBreakHandle.java
@@ -18,7 +18,7 @@ public class NotBreakHandle implements Handler {
 	private static final Logger logger = LoggerFactory.getLogger(NotBreakHandle.class);
 
 	@Override
-	public boolean handler(Sender sender, int clientId, Message message, int mapId, int sequence) {
+	public boolean handler(Sender sender, int clientId, Message message, long mapId, int sequence) {
 		try {
 			ModelProto.NotBreak notification = (ModelProto.NotBreak) message;
 			String clientIp = notification.getCert().toStringUtf8();
