+++ center/src/main/java/center/client/handle/NotClientLinkHandle.java
@@ -67,7 +67,7 @@ public class NotClientLinkHandle implements Handler {
 	}
 
 	@Override
-	public boolean handler(Sender sender, int clientId, Message message, int mapId, int sequence) {
+	public boolean handler(Sender sender, int clientId, Message message, long mapId, int sequence) {
 		try {
 			ModelProto.NotRegisterClient notification = (ModelProto.NotRegisterClient) message;
 			String clientIp = notification.getCert().toStringUtf8();
