+++ room/src/main/java/room/client/handle/server/gate/NotBreakHandle.java
@@ -20,7 +20,7 @@ public class NotBreakHandle implements Handler {
 	private static final Logger logger = LoggerFactory.getLogger(NotBreakHandle.class);
 
 	@Override
-	public boolean handler(Sender sender, int clientId, Message msg, int mapId, int sequence) {
+	public boolean handler(Sender sender, int clientId, Message msg, long mapId, int sequence) {
 		try {
 			ModelProto.NotBreak notice = (ModelProto.NotBreak) msg;
 			int userId = notice.getUserId();
